Added undo_excavation command Added partial German translations (ACGaming) Added support for mouse side buttons Changed excavate key polling slightly Minor optimisations to mining passes Minor bug fix to item checks

+++ src/main/java/oreexcavation/utils/XPHelper.java
@@ -17,9 +17,9 @@ public class XPHelper
 		}
 	}
 	
-	public static void addXP(EntityPlayer player, int xp, boolean sync)
+	public static void addXP(EntityPlayer player, long xp, boolean sync)
 	{
-		long experience = getPlayerXP(player) + (long)xp;
+		long experience = getPlayerXP(player) + xp;
 		player.experienceTotal = experience >= Integer.MAX_VALUE? Integer.MAX_VALUE : (int)experience;
 		player.experienceLevel = getXPLevel(experience);
 		long expForLevel = getLevelXP(player.experienceLevel);
